Add 'Coding' suffix to the class names of coding challenges
