+++ knpu/Challenges/FixMissingUseStatementCoding.php
@@ -11,15 +11,17 @@ use KnpU\Gladiator\Grading\HtmlOutputGradingTool;
 use KnpU\Gladiator\Grading\PhpGradingTool;
 use KnpU\Gladiator\Worker\WorkerLoaderInterface;
 
-class AddAliasToUseStatement implements CodingChallengeInterface
+class FixMissingUseStatementCoding implements CodingChallengeInterface
 {
     public function getQuestion()
     {
         return <<<EOF
-The intern is messing around and has changed the `use`
-statement of alias `Pizza` to `ItalianTreat`. Without
-changing that line, update the `new` line so that the
-code runs using the new alias.
+After lunch, the whole team is happily coding with
+namespaces and `use` statements. Suddenly, someone
+sees an error and calls you over for help! Can you
+fix this code?
+
+P.S. This is the *most* common mistake when using namespaces.
 EOF;
 
     }
@@ -29,18 +31,20 @@ EOF;
         $builder = new ChallengeBuilder();
 
         $builder
-            ->addFileContents('eat.php', <<<EOF
+            ->addFileContents('Lunch.php', <<<EOF
 <?php
 
-require 'Pizza.php';
+namespace Meals;
 
-use Food\Tasty\Pizza as ItalianTreat;
-
-\$pizza = new Pizza();
-
-echo \$pizza->eat();
+class Lunch
+{
+    public function getFood()
+    {
+        return new Pizza();
+    }
+}
 EOF
-        )
+            )
             ->addFileContents('Pizza.php', <<<EOF
 <?php
 
@@ -54,7 +58,19 @@ class Pizza
     }
 }
 EOF
-        , true)
+            , true)
+            ->addFileContents('eat.php', <<<EOF
+<?php
+
+require 'Pizza.php';
+require 'Lunch.php';
+
+use Meals\Lunch;
+
+\$lunch = new Lunch();
+echo \$lunch->getFood()->eat();
+EOF
+            , true)
             ->setEntryPointFilename('eat.php')
         ;
 
@@ -77,30 +93,39 @@ EOF
 
         $htmlGrader->assertOutputContains('COWP', 'The output no longer has the COWP! Is something wrong?');
         $phpGrader->assertInputContains(
-            'eat.php',
-            'use Food\Tasty\Pizza as ItalianTreat;',
-            'Don\'t change a `use` line with the `ItalianTreat` alias. Update a line where creating new instance of the `Pizza` class instead.'
+            'Lunch.php',
+            'return new Pizza',
+            'Try to fix it without changing the `Pizza` class name in `getFood()` method.'
+        );
+        $phpGrader->assertInputDoesNotContain(
+            'Lunch.php',
+            'use \Food\Tasty\Pizza;',
+            'Don\'t use `\\` at the beginning of namespace in `use` statement. It doesn\'t necessary.'
         );
         $phpGrader->assertInputContains(
-            'eat.php',
-            'new ItalianTreat',
-            'Did you use the `ItalianTreat` alias for creating an instance of the `Pizza` class?'
+            'Lunch.php',
+            'use Food\Tasty\Pizza;',
+            'Seems we need to add a `use` statement for the `Pizza` class in the `Lunch.php` file.'
         );
     }
 
     public function configureCorrectAnswer(CorrectAnswer $correctAnswer)
     {
         $correctAnswer
-            ->setFileContents('eat.php', <<<EOF
+            ->setFileContents('Lunch.php', <<<EOF
 <?php
 
-require 'Pizza.php';
+namespace Meals;
 
-use Food\Tasty\Pizza as ItalianTreat;
+use Food\Tasty\Pizza;
 
-\$pizza = new ItalianTreat();
-
-echo \$pizza->eat();
+class Lunch
+{
+    public function getFood()
+    {
+        return new Pizza();
+    }
+}
 EOF
             )
         ;
